requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-vision
google-auth
pydantic
